api-gw/pyproject.toml
[tool.poetry]
name = "api-gw"
version = "0.1.0"
description = ""
authors = ["Eduard Badillo <eduard.badillo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
opentelemetry-exporter-otlp = "^1.16.0"
opentelemetry-api = "^1.16.0"
opentelemetry-sdk = "^1.16.0"
opentelemetry-instrumentation-fastapi = "^0.37b0"
requests = "^2.28.2"
prometheus-fastapi-instrumentator = "^5.9.1"
fastapi = "0.90.1"
aiohttp = "^3.8.4"
opentelemetry-instrumentation-aiohttp-client = "^0.37b0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
